requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.3.10
groq==0.5.0
python-dotenv==1.0.1
